agent: update server and add client implementation
This commit update the server agent based on the SSH protocol, adding an extended implementation to manage the CLI sessions along with the SSH keyring available via Golang SSH agent implementation.

Additionally a first implementation of a client for the agent that is capable to handle the CLI session has been implemented.